Sync develop branch with EMC develop

+++ ush/wave_tar.sh
@@ -1,4 +1,5 @@
-#!/bin/bash
+#! /usr/bin/env bash
+
 ###############################################################################
 #                                                                             #
 # This script tars the sectral or bulletin files into a single file and       #
@@ -23,19 +24,15 @@
 #
 # --------------------------------------------------------------------------- #
 # 0.  Preparations
+
+source "$HOMEgfs/ush/preamble.sh"
+
 # 0.a Basic modes of operation
 
-  # set execution trace prompt.  ${0##*/} adds the script's basename
-  PS4=" \${SECONDS} ${0##*/} L\${LINENO} + "
-  set -x
-
-  # Use LOUD variable to turn on/off trace.  Defaults to YES (on).
-  export LOUD=${LOUD:-YES}; [[ $LOUD = yes ]] && export LOUD=YES
-  [[ "$LOUD" != YES ]] && set -x
-
   cd $DATA
-  postmsg "$jlogfile" "Making TAR FILE"
+  echo "Making TAR FILE"
 
+  alertName=$(echo $RUN|tr [a-z] [A-Z])
 
   set +x
   echo ' '
@@ -45,7 +42,7 @@
   echo "   ID              : $1"
   echo "   Type            : $2"
   echo "   Number of files : $3"
-  [[ "$LOUD" = YES ]] && set -x
+  set_trace
 
 
 # 0.b Check if type set
@@ -58,8 +55,7 @@
     echo '*** VARIABLES IN ww3_tar.sh NOT SET ***'
     echo '********************************************'
     echo ' '
-    [[ "$LOUD" = YES ]] && set -x
-    postmsg "$jlogfile" "TYPE IN ww3_tar.sh NOT SET"
+    set_trace
     exit 1
   else
     ID=$1
@@ -69,6 +65,9 @@
 
   filext=$type
   if [ "$type" = "ibp" ]; then filext='spec'; fi
+  if [ "$type" = "ibpbull" ]; then filext='bull'; fi
+  if [ "$type" = "ibpcbull" ]; then filext='cbull'; fi
+
 
   rm -rf TAR_${filext}_$ID 
   mkdir  TAR_${filext}_$ID
@@ -77,21 +76,19 @@
 # 0.c Define directories and the search path.
 #     The tested variables should be exported by the postprocessor script.
 
-  if [ -z "$cycle" ] || [ -z "$COMOUT" ] || [ -z "$WAV_MOD_TAG" ] ||  \
-     [ -z "$SENDCOM" ] || [ -z "$SENDDBN" ] || [ -z "${STA_DIR}" ]
-  then
+  if [[ -z "${cycle}" ]] || [[ -z "${COM_WAVE_STATION}" ]] || [[ -z "${WAV_MOD_TAG}" ]] ||  \
+     [[ -z "${SENDDBN}" ]] || [[ -z "${STA_DIR}" ]]; then
     set +x
     echo ' '
     echo '*****************************************************'
     echo '*** EXPORTED VARIABLES IN ww3_tar.sh NOT SET ***'
     echo '*****************************************************'
     echo ' '
-    [[ "$LOUD" = YES ]] && set -x
-    postmsg "$jlogfile" "EXPORTED VARIABLES IN ww3_tar.sh NOT SET"
+    set_trace
     exit 2
   fi
 
-  cd ${STA_DIR}/${type}
+  cd ${STA_DIR}/${filext}
 
 # --------------------------------------------------------------------------- #
 # 2.  Generate tar file (spectral files are compressed)
@@ -99,7 +96,7 @@
   set +x
   echo ' '
   echo '   Making tar file ...'
-  set -x
+  set_trace
 
   count=0
   countMAX=5
@@ -108,7 +105,7 @@
   while [ "$count" -lt "$countMAX" ] && [ "$tardone" = 'no' ]
   do
     
-    nf=`ls | awk '/'$ID.*.$filext'/ {a++} END {print a}'`
+    nf=$(ls | awk '/'$ID.*.$filext'/ {a++} END {print a}')
     nbm2=$(( $nb - 2 ))
     if [ $nf -ge $nbm2 ]
     then 
@@ -123,8 +120,7 @@
         echo '*** FATAL ERROR : TAR CREATION FAILED *** '
         echo '***************************************** '
         echo ' '
-        [[ "$LOUD" = YES ]] && set -x
-        postmsg "$jlogfile" "FATAL ERROR : TAR CREATION FAILED"
+        set_trace
         exit 3
       fi
       
@@ -135,9 +131,9 @@
     else
       set +x
       echo ' All files not found for tar. Sleeping 10 seconds and trying again ..'
-      [[ "$LOUD" = YES ]] && set -x
+      set_trace
       sleep 10
-      count=`expr $count + 1`
+      count=$(expr $count + 1)
     fi
 
   done
@@ -150,12 +146,11 @@
     echo '*** FATAL ERROR : TAR CREATION FAILED *** '
     echo '***************************************** '
     echo ' '
-    [[ "$LOUD" = YES ]] && set -x
-    postmsg "$jlogfile" "FATAL ERROR : TAR CREATION FAILED"
+    set_trace
     exit 3
   fi
 
-  if [ "$filext" = 'spec' ]
+  if [ "$type" = 'spec' ]
   then
     if [ -s $ID.$cycle.${type}_tar ]
     then
@@ -171,8 +166,7 @@
         echo '*** FATAL ERROR : SPECTRAL TAR COMPRESSION FAILED *** '
         echo '***************************************************** '
         echo ' '
-        [[ "$LOUD" = YES ]] && set -x
-        postmsg "$jlogfile" "FATAL ERROR : SPECTRAL TAR COMPRESSION FAILED"
+        set_trace
         exit 4
       fi
     fi
@@ -185,10 +179,10 @@
 
   set +x
   echo ' '
-  echo "   Moving tar file ${file_name} to $COMOUT ..."
-  [[ "$LOUD" = YES ]] && set -x
+  echo "   Moving tar file ${file_name} to ${COM_WAVE_STATION} ..."
+  set_trace
 
-  cp ${file_name} $COMOUT/station/.
+  cp "${file_name}" "${COM_WAVE_STATION}/."
 
   exit=$?
 
@@ -200,8 +194,7 @@
     echo '*** FATAL ERROR : TAR COPY FAILED *** '
     echo '************************************* '
     echo ' '
-    [[ "$LOUD" = YES ]] && set -x
-    postmsg "$jlogfile" "FATAL ERROR : TAR COPY FAILED"
+    set_trace
     exit 4
   fi
 
@@ -209,23 +202,22 @@
   then
     set +x
     echo ' '
-    echo "   Alerting TAR file as $COMOUT/station/${file_name}"
+    echo "   Alerting TAR file as ${COM_WAVE_STATION}/${file_name}"
     echo ' '
-    [[ "$LOUD" = YES ]] && set -x
-    $DBNROOT/bin/dbn_alert MODEL OMBWAVE $job $COMOUT/station/${file_name}
+    set_trace
+    "${DBNROOT}/bin/dbn_alert MODEL" "${alertName}_WAVE_TAR" "${job}" \
+      "${COM_WAVE_STATION}/${file_name}"
   fi
 
 # --------------------------------------------------------------------------- #
 # 4.  Final clean up
 
-  cd $DATA
+cd $DATA
 
-  set +x; [[ "$LOUD" = YES ]] && set -v
+if [[ ${KEEPDATA:-NO} == "NO" ]]; then
+  set -v
   rm -rf  ${STA_DIR}/${type}
   set +v
-
-  echo ' '
-  echo 'End of ww3_tar.sh at'
-  date
+fi
 
 # End of ww3_tar.sh ----------------------------------------------------- #
